feat:fixing of the prefinal fixing of CARF
