+++ src/pages/api/survey/index.ts
@@ -2,7 +2,6 @@ import { NextApiRequest, NextApiResponse } from "next";
 import { PrismaClient } from "@prisma/client";
 import { getServerSession } from "next-auth/next";
 import { authOptions } from "@/pages/api/auth/[...nextauth]";
-import Email from "next-auth/providers/email";
 
 const prisma = new PrismaClient();
 
@@ -15,25 +14,12 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
     }
     console.log(session);
     const { name } = req.body;
-    const existingUser = await prisma.user.findUnique({
-      where: {
-        id: session.user.id,
-      },
-    });
-
-    if (!existingUser) {
-      return res.status(400).json({ error: "Invalid userId" });
-    }
 
     try {
       const survey = await prisma.survey.create({
         data: {
           name,
-          user: {
-            connect: {
-              id: session.user.id,
-            },
-          },
+          userId: session.user.id,
         },
       });
